+++ crates/radiate-engines/src/context.rs
@@ -83,7 +83,7 @@ impl<C: Chromosome, T> EvolutionContext<C, T> {
 
         self.metrics
             .replace(metric!(metric_names::INDEX, self.index));
-        self.metrics.bump(self.index as u64);
+        self.metrics.bump(self.index);
 
         Ok(best_improved)
     }
